Move `BindingDoc` into new private package.
The goal is to move the documentation model POJOs into a package used by both the runtime and the HTML generation tool.
