+++ src/main/java/dev/ionfusion/fusion/ModuleNamespace.java
@@ -7,6 +7,7 @@ import static dev.ionfusion.fusion.BindingSite.makeExportBindingSite;
 import static dev.ionfusion.fusion.GlobalState.REQUIRE;
 
 import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
+import dev.ionfusion.fusion._private.doc.model.BindingDoc;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Iterator;
