+++ src/main/java/dev/ionfusion/fusion/ModuleStore.java
@@ -4,6 +4,7 @@
 package dev.ionfusion.fusion;
 
 import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
+import dev.ionfusion.fusion._private.doc.model.BindingDoc;
 
 /**
  * Eval-time storage for modules.
